requirements.txt
asgiref==3.3.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
Django==3.1.7
djangorestframework==3.12.2
Flask==1.0.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.1.1
pytz==2021.1
requests==2.18.4
sqlparse==0.4.1
urllib3==1.22
Werkzeug==0.14.1
